+++ src/queue/pending_queue.rs
@@ -110,4 +110,11 @@ impl PendingQueue {
     pub(crate) fn is_empty(&self) -> bool {
         self.len() == 0
     }
+
+    pub(crate) fn contains_stream(&self, stream_identifier: u16) -> bool {
+        self.unordered_queue
+            .iter()
+            .chain(self.ordered_queue.iter())
+            .any(|chunk| chunk.stream_identifier == stream_identifier)
+    }
 }